cleanup spark transformer package name, rename HiveUtil -> SparkSQLUtil
