+++ sdl-spark/src/main/scala/io/smartdatalake/workflow/action/spark/transformer/DecryptColumnsTransformer.scala
@@ -1,5 +1,5 @@
 /*
- * Smart Data Lake Builder - Build your data lake the smart way.
+ * Smart Data Lake - Build your data lake the smart way.
  *
  * Copyright © 2019-2026 ELCA Informatique SA (<https://www.elca.ch>)
  *
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package io.smartdatalake.workflow.action.generic.transformer
+package io.smartdatalake.workflow.action.spark.transformer
 
 import com.typesafe.config.Config
 import io.smartdatalake.config.SdlConfigObject.{ActionId, DataObjectId}
@@ -25,22 +25,23 @@ import io.smartdatalake.util.crypt.{EncryptDecrypt, EncryptDecryptECB, EncryptDe
 import io.smartdatalake.util.hdfs.PartitionValues
 import io.smartdatalake.util.secrets.StringOrSecret
 import io.smartdatalake.workflow.ActionPipelineContext
+import io.smartdatalake.workflow.action.generic.transformer.GenericDfTransformer
 import org.apache.spark.sql.DataFrame
 
 /**
- * Encryption of specified columns using AES/GCM algorithm.
+ * Decryption of specified columns using AES/GCM algorithm.
  *
  * @param name           name of the transformer
  * @param description    Optional description of the transformer
- * @param encryptColumns List of columns [columnA, columnB] to be encrypted
+ * @param decryptColumns List of columns [columnA, columnB] to be encrypted
  * @param key            contains the id of the provider and the name of the secret with format <PROVIDERID>#<SECRETNAME>,
  *                       e.g. ENV#<ENV_VARIABLE_NAME> to get a secret from an environment variable OR CLEAR#mYsEcReTkeY
  * @param algorithm      Specify: "GCM" (AES/GCM/NoPadding), "ECB" (AES/ECB/PKCS5Padding),
  *                       alternatively a class name extending trait EncryptDecrypt can be provided. DEFAULT: GCM
  */
-case class EncryptColumnsTransformer(override val name: String = "encryptColumns",
+case class DecryptColumnsTransformer(override val name: String = "encryptColumns",
                                      override val description: Option[String] = None,
-                                     encryptColumns: Seq[String],
+                                     decryptColumns: Seq[String],
                                      private val key: StringOrSecret,
                                      algorithm: String = "GCM"
                                     )
@@ -54,15 +55,16 @@ case class EncryptColumnsTransformer(override val name: String = "encryptColumns
     case _ => throw new UnsupportedOperationException(s"unsupported en/decryption algorithm $algorithm")
   }
 
-  override def transform(actionId: ActionId, partitionValues: Seq[PartitionValues], df: DataFrame, dataObjectId: DataObjectId)(implicit context: ActionPipelineContext): DataFrame = {
-    crypt.encryptColumns(df, encryptColumns)
+  override def transform(actionId: ActionId, partitionValues: Seq[PartitionValues],
+                         df: DataFrame, dataObjectId: DataObjectId)(implicit context: ActionPipelineContext): DataFrame = {
+    crypt.decryptColumns(df, decryptColumns)
   }
 
-  override def factory: FromConfigFactory[GenericDfTransformer] = EncryptColumnsTransformer
+  override def factory: FromConfigFactory[GenericDfTransformer] = DecryptColumnsTransformer
 }
 
-object EncryptColumnsTransformer extends FromConfigFactory[GenericDfTransformer] {
-  override def fromConfig(config: Config)(implicit instanceRegistry: InstanceRegistry): EncryptColumnsTransformer = {
-    extract[EncryptColumnsTransformer](config)
+object DecryptColumnsTransformer extends FromConfigFactory[GenericDfTransformer] {
+  override def fromConfig(config: Config)(implicit instanceRegistry: InstanceRegistry): DecryptColumnsTransformer = {
+    extract[DecryptColumnsTransformer](config)
   }
 }
