+++ sdl-spark/src/main/scala/io/smartdatalake/workflow/action/spark/transformer/PythonCodeSparkDfsTransformer.scala
@@ -25,7 +25,7 @@ import io.smartdatalake.util.hdfs.{HdfsUtil, PartitionValues}
 import io.smartdatalake.util.misc.DefaultExpressionData
 import io.smartdatalake.util.spark.{PythonSparkEntryPoint, PythonUtil}
 import io.smartdatalake.workflow.ActionPipelineContext
-import io.smartdatalake.workflow.action.generic.transformer.{GenericDfsTransformer, OptionsSparkDfsTransformer}
+import io.smartdatalake.workflow.action.generic.transformer.GenericDfsTransformer
 import io.smartdatalake.workflow.action.spark.transformer.PythonCodeDfTransformer.dedent
 import io.smartdatalake.workflow.dataframe.spark.SparkSubFeed.getSparkSession
 import org.apache.hadoop.conf.Configuration
