+++ sdl-spark/src/main/scala/io/smartdatalake/workflow/action/spark/transformer/ScalaNotebookSparkDfTransformer.scala
@@ -25,7 +25,7 @@ import io.smartdatalake.util.hdfs.PartitionValues
 import io.smartdatalake.util.misc.{CustomCodeUtil, DefaultExpressionData}
 import io.smartdatalake.util.webservice.SttpWebserviceClient
 import io.smartdatalake.workflow.ActionPipelineContext
-import io.smartdatalake.workflow.action.generic.transformer.{GenericDfTransformer, OptionsSparkDfTransformer}
+import io.smartdatalake.workflow.action.generic.transformer.GenericDfTransformer
 import io.smartdatalake.workflow.action.spark.customlogic.CustomDfTransformerConfig.fnTransformType
 import io.smartdatalake.workflow.connection.authMode.AuthMode
 import io.smartdatalake.workflow.dataframe.spark.SparkSubFeed.getSparkSession
